+++ README.md
@@ -1,15 +1,2 @@
 # CaC-23545_TP_Final
 Trabajo Final del curso Full Stack Java de Codo a Codo
-
-# Requisitos
-Para ejecutar el proyecto necesitas crear el archivo `config.properties` dentro la carpeta `src/main/resources/`.
-
-Los datos que debe contener el archivo de configuracion deben ser los siguientes:
-
-```config
-db.host=
-db.username=
-db.password=
-db.port=
-db.name=
-```
