+++ Servers/.settings/org.eclipse.wst.server.core.prefs
@@ -1,0 +1,3 @@
+#Wed Nov 29 17:10:57 ART 2023
+org.eclipse.wst.server.core.isServerProject=true
+eclipse.preferences.version=1
